Implement proper QE Report signature and attestation key verification
This commit fixes critical security gaps in the attestation key certification
verification for both C and Go implementations.

Previously, verify_attestation_key_internal only checked that both keys were
EC keys, which provided no security. An attacker could substitute their own
attestation key and generate fake quotes.

Changes:
- Fix cert_data_size field to uint32 (was incorrectly uint16), which caused
  the leaf PCK certificate to be skipped during parsing
- Implement verify_qe_report_signature() to verify the QE Report is signed
  by the PCK certificate's public key
- Implement verify_qe_report_data() to verify the attestation key hash is
  bound to the QE Report's report_data field
- Update verify_attestation_key_internal() to perform proper cryptographic
  verification instead of just type checking
- Add VerifyAttestationKeyCertification() to Go implementation with same
  verification logic
- Update tests to expect correct 3-certificate chain (leaf PCK, intermediate
  CA, root CA)

The verification now properly establishes the trust chain:
  Quote → Attestation Key → QE Report → PCK Certificate → Intel Root CA

+++ src/include/sgx_types.h
@@ -133,7 +133,7 @@ typedef struct _sgx_ql_auth_data_t {
     uint16_t              auth_data_size;       /* Size of auth data (observed: 0x0020) */
     uint8_t               auth_data[32];        /* 32 bytes of auth data (0x00-0x1F) */
     uint16_t              cert_type;            /* Certificate type (observed: 0x0005) */
-    uint16_t              cert_data_size;       /* Size of certificate data (observed: ~0x0DDC) */
+    uint32_t              cert_data_size;       /* Size of certificate data (4 bytes, observed: ~0x0DDC) */
     uint8_t               cert_data[];          /* Certificate data (PEM formatted PCK certs) */
 } sgx_ql_auth_data_t;
